refacor: 코드 리뷰에 맞춰 빠진부분 보완 및 수정, 함수명 변경
