+++ MoodStation/MoodStation/Home/HomeView.swift
@@ -42,7 +42,7 @@ final class HomeView: UIView {
         self.tableView.do {
             $0.backgroundColor = .customBlack
             $0.separatorStyle = .none
-            $0.register(RecordListCell.self, forCellReuseIdentifier: RecordListCell.className)
+            $0.register(RecordListCell.self)
         }
     }
