+++ MoodStation/MoodStation/Settings/SettingsViewController.swift
@@ -114,7 +114,9 @@ extension SettingsViewController: SettingsViewDataSource {
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let model = self.viewModel.cellModel(at: indexPath)
+        guard let model = self.viewModel.cellModel(at: indexPath) else {
+            return UITableViewCell()
+        }
         
         switch model {
         case .text(let textOnlyTableViewCellModel):
